# services/route-service/internal/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves as an interface to interact with the My Route Service platform, providing endpoints for managing bus routes, Stops, and schedules.",
        "title": "My Route Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Mohammad Mizan",
            "url": "http://swagger.io/support",
            "email": "takbir.jcd@gmail.com"
        },
        "license": {
            "name": "Apache License Version 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/routes/stops/{stopId}/schedules": {
            "get": {
                "description": "Retrieve all schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ScheduleResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid route ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new schedule for a specific route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "stopId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule information",
                        "name": "addScheduleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created schedule details",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route ID, request format, or validation error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/routes/stops/{stopId}/schedules/{id}": {
            "get": {
                "description": "Retrieve a schedule by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Stop ID",
                        "name": "stopId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule details",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid schedule ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of a schedule by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated schedule information",
                        "name": "updateScheduleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated schedule details",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid schedule ID or request format",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a schedule by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid schedule ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Schedule not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddScheduleRequest": {
            "type": "object",
            "required": [
                "arrival_time",
                "departure_time",
                "stop_id"
            ],
            "properties": {
                "arrival_time": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "stop_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "arrival_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateScheduleRequest": {
            "type": "object",
            "required": [
                "arrival_time",
                "departure_time",
                "stop_id"
            ],
            "properties": {
                "arrival_time": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "stop_id": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}